User guide

Introduction

We start off by giving some mathematical background, or rather by defining the needed language.

Weakly separated Collections

For any integer $n \geq 1$ we use the notation $[n]:=$ { $1, 2, \ldots, n$ } and denote by $\text{Pot}(k,n)$ the set of $k$-subsets of $[n]$.

Definition (weak separation)

Let $I, J$ be $k$-subsets of $[n]$, then we call $I$ and $J$ $\textbf{weakly separated}$ if we cannot find elements $a, c \in I \setminus J$ and $b, d \in J \setminus I$ such that $(a, b, c, d)$ is strictly cyclically ordered. In this case we write $I \parallel J$.

Intuitively two $k$-subsets are weakly separated if after can arranging $I \setminus J$ and $J \setminus I$ clockwise on a circle, they can be separated by a line.

Definition (weakly separated collection)

A subset $\mathcal{C} \subseteq \text{Pot}(k,n)$ is called a $\textbf{weakly separated collection}$ (abbreviated by WSC) if its elements are pairwise weakly separated. We often referr to elements of a WSC as labels.

Definition (mutation)

If $\mathcal{C}$ is a WSC that includes sets of the form $Iab, Ibc, Icd, Iad$ and $Iac$, where $(a,b,c,d)$ is strictly cyclically ordered. Then $\mathcal{C'} = (\mathcal{C} \setminus \{Iac\}) \cup \{Ibd\}$ is also a weakly separated collection, and we call the exchange of $Iac$ by $Ibd$ a $\textbf{mutation}$.

Plabic Tilings

Definition (plabic tiling)

Any WSC can be given the structure of an abstract $2$-dimensional cell complex, which in turn may be embedded into the plane. This construction will be called an (abstract) $\textbf{plabic tiling}$, and we referr to TODO for the mathematical details.

Intuitively a plabic tiling is a tiling of a convex $n$-gon into convex polygons, colored either black or white, and with vertices labelled by the elements of the underlying WSC. Plabic tilings are in bijective correspondance with WSC's that are maximal with respect to inclusion.

Plabic Graphs

Definition (plabic graph)

A $\textbf{plabic graph}$ is a finite simple connected plane graph $G$ whose interior is bounded by a vertex disjoint cycle containing $n$ $\textbf{boundary vertices}$ $b_1, \ldots, b_n$. Here the labelling is chosen in clockwise order.

We only consider $\textbf{reduced}$ plabic graphs which can be also seen to be in one to one correspondance to WSC's that are maximal with respect to inclusion. For more details we referr to TODO.

Combinatorics

In this section we will learn how to use the combinatorial part of WeaklySeparatedCollections.

Vectors instead of sets

In this package we use vecors in place of sets, although WSC's are by definition sets of $k$-sets. We always assume such vectors to be increasingly ordered and not contain double elements. None of the below methods check these properties and unforseen behavior may arise if they are not fulfilled.

To see if two or more $k$-subsets are weakly separated, we use the function is_weakly_separated.

WeaklySeparatedCollections.is_weakly_separated — Function

is_weakly_separated(n::Int, v::Vector{Int}, w::Vector{Int})

Test if two vectors v and w viewed as subsets of {1 , ..., n } are weakly separated.

Examples

julia> v = [1,2,3,5,6,9]
julia> w = [1,2,4,5,7,8]
julia> is_weakly_separated(9, v, w)
true

julia> v = [1,2,3,5,6,9]
julia> w = [1,2,3,5,7,8]
julia> is_weakly_separated(9, v, w)
false

is_weakly_separated(n::Int, labels::Vector{Vector{Int}})

Test if the vectors contained in labels are pairwise weakly separated.

Examples

julia> u = [1,2,3,4,5,6]
julia> v = [1,2,3,5,6,9]
julia> w = [1,2,3,5,7,8]
julia> is_weakly_separated(9, [u, v, w])
true

The labels of some known weakly separated collections are available via

WeaklySeparatedCollections.rectangle_labels — Function

rectangle_labels(k::Int, n::Int)

Return the labels of the rectangle graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.checkboard_labels — Function

checkboard_labels(k::Int, n::Int)

Return the labels of the checkboard graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.dual_rectangle_labels — Function

dual_rectangle_labels(k::Int, n::Int)

Return the labels of the dual-rectangle graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.dual_checkboard_labels — Function

dual_checkboard_labels(k::Int, n::Int)

Return the labels of the dual-checkboard graph as a vector. The frozen labels are in positions 1 to n.

The data type for WSC's or rather (abstract) plabic tilings is given by WSCollection.

WeaklySeparatedCollections.WSCollection — Type

WSCollection

An abstract 2-dimensional cell complex living inside the matriod of k-sets in {1, ..., n}. Its vertices are labelled by elements of labels while quiver encodes adjacencies between the vertices. The 2-cells are colored black or white and contained in blackCliques and whiteCliques.

Optionally the 2-cells can be set to missing, to save memory.

Attributes

k::Int

n::Int

labels::Vector{Vector{Int}}

quiver::SimpleDiGraph{Int}

whiteCliques::Union{Missing, Dict{Vector{Int}, Vector{Int} }}

blackCliques::Union{Missing, Dict{Vector{Int}, Vector{Int} }}

Constructors

WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, computeCliques::Bool = true)
WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, quiver::SimpleDiGraph{Int}, 
                                                          computeCliques::Bool = true)

Examples

julia> labels = rectangle_labels(4, 9)
julia> WSCollection(4, 9, labels);

The above predefined sets of labels give rise to weakly separated collections which are available via

WeaklySeparatedCollections.checkboard_collection — Function

checkboard_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the checkboard graph.

WeaklySeparatedCollections.rectangle_collection — Function

rectangle_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the rectangle graph.

WeaklySeparatedCollections.dual_checkboard_collection — Function

dual_checkboard_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the dual-checkboard graph.

WeaklySeparatedCollections.dual_rectangle_collection — Function

dual_rectangle_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the dual-rectangle graph.

We may easily compare two WSC's

Missing docstring.

Missing docstring for isequal. Check Documenter's build log for details.

Base.:== — Function

(==)(collection1::WSCollection, collection2::WSCollection)

Return true if the vertices of collection1 and collection2 share the same labels. The order of labels in each collection does not matter.

WSC's usually contain frozen elements that never change. On the other hand some elements may be modified via mutation and are called mutable. To figure out which elemnents of a WSC are frozen or mutable use the functions is_frozen or is_mutable.

WeaklySeparatedCollections.is_frozen — Function

is_frozen(collection::WSCollection, i::Int)

Return true if the vertex i of collection is frozen.

Examples

julia> H = rectangle_collection(4, 9)
julia> is_frozen(H, 5)
true

julia> is_frozen(H, 11)
false

WeaklySeparatedCollections.is_mutable — Function

is_mutable(collection::WSCollection, i::Int)

Return true if the vertex i of collection is mutable. This is the case if it is not frozen and is of degree 4.

Examples

julia> H = rectangle_collection(4, 9)
julia> is_mutable(H, 11)
false

julia> is_frozen(H, 10)
true

To mutate a WSC, the functions mutateand mutate! are available.

WeaklySeparatedCollections.mutate! — Function

mutate!(collection::WSCollection, i::Int, mutateCliques::Bool = true)

Mutate the collection in direction i if i is a mutable vertex of collection.

If mutateCliques is set to false, the 2-cells are set to missing.

Examples

julia> H = rectangle_collection(4, 9)
julia> mutate!(H, 10)

mutate!(collection::WSCollection, label::Vector{Int}, mutateCliques::Bool = true)

Mutate the collection by addressing a vertex with its label.

Examples

julia> H = rectangle_collection(4, 9)
julia> H.labels[10]
4-element Vector{Int64}:
 2
 7
 8
 9

julia> mutate!(H, [2,7,8,9])

WeaklySeparatedCollections.mutate — Function

mutate(collection::WSCollection, i::Int, mutateCliques::Bool = true)

Version of mutate! that does not modify its arguments.

mutate(collection::WSCollection, label::Vector{Int}, mutateCliques::Bool = true)

Mutate the collection by addressing a vertex with its label, without modifying arguments.

Apart from mutation, several other transformations of WSC's are available:

Missing docstring.

Missing docstring for rotate_collection!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for rotate_collection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reflect_collection!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for reflect_collection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for complement_collection!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for complement_collection. Check Documenter's build log for details.

Missing docstring.

Missing docstring for swaped_colors_collection!. Check Documenter's build log for details.

Missing docstring.

Missing docstring for swaped_colors_collection. Check Documenter's build log for details.

We often want to deal with maximal WSC's instead of their subsets. To extend a given WSC to a maximal one, the following functions may be used:

WeaklySeparatedCollections.extend_weakly_separated! — Function

extend_weakly_separated!(k::Int, n::Int, labels::Vector{Vector{Int}})

Extend labels to contain the labels of a maximal weakly separated collection.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> extend_weakly_separated!(4, 9, labels)

extend_weakly_separated!(k::Int, n::Int, labels1::Vector{Vector{Int}}, 
                                         labels2::Vector{Vector{Int}})

Extend labels1 to contain the labels of a maximal weakly separated collection. Use elements of labels2 if possible.

Examples

julia> labels1 = [[1,3,5,7], [1,3,5,9]]
julia> labels2 = checkboard_labels(4, 9)
julia> extend_weakly_separated!(4, 9, labels1, labels2)

extend_weakly_separated!(labels::Vector{Vector{Int}}, collection::WSCollection)

Extend labels to contain the labels of a maximal weakly separated collection. Use labels of collection if possible.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> H = checkboard_collection(4, 9)
julia> extend_weakly_separated!(labels, H)

WeaklySeparatedCollections.extend_to_collection — Function

extend_to_collection(k::Int, n::Int, labels::Vector{Vector{Int}})

Return a maximal weakly separated collection containing all elements of labels.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> extend_to_collection(4, 9, labels)

extend_to_collection(k::Int, n::Int, labels1::Vector{Vector{Int}}, 
                                     labels2::Vector{Vector{Int}})

Return a maximal weakly separated collection containing all elements of labels1. Use elements of labels2 if possible.

Examples

julia> labels1 = [[1,3,5,7], [1,3,5,9]]
julia> labels2 = checkboard_labels(4, 9)
julia> extend_to_collection(4, 9, labels1, labels2)

extend_to_collection(labels::Vector{Vector{Int}}, collection::WSCollection)

Return a maximal weakly separated collection containing all elements of labels. Use labels of collection if possible.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> H = checkboard_collection(4, 9)
julia> extend_to_collection(labels, H)

Plotting

Plotting WSC's requires Luxor to be installed and loaded as detailed here.

In the introduction we learned about plabic tilings as well as plabic graphs as objects living in the plane which are in one to one correspndance to maximal WSC's. Thus we can plot a maximal WSC using its corresponding plabic tiling or plabic graph. The functions to accomplish this are:

WeaklySeparatedCollections.drawTiling — Function

drawTiling(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic tiling of the provided weakly separated collection and save it as an image file of specified size. Both the name as well as the resulting file type of the image are controlled via title.

Inside a Jupyter sheet drawing without saving an image is possible by omitting the title argument i.e. via drawTiling(collection::WSCollection, width::Int = 500, height::Int = 500).

Keyword Arguments

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

adjustAngle::Bool = false

highlightMutables::Bool = true

labelDirection = "left"

labelDirection controls whether labels the "left" (i.e. the usual ones) or "right" (complements) labels are drawn.

Missing docstring.

Missing docstring for drawPLG_straight. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drawPLG_smooth. Check Documenter's build log for details.

Missing docstring.

Missing docstring for drawPLG_poly. Check Documenter's build log for details.

Graphical user interface

This section is work in progress.

The graphical user interface requires both an installation of Luxor as well as Mousetrap. See here for details.

While plotting WSC's enables us to visualize them, the resulting images lack interactivity. This is where the built in gui application comes in handy. To start it we use

WeaklySeparatedCollections.visualizer! — Function

visualizer!(collection::WSCollection = rectangle_collection(4, 9))

Start the graphical user interface to visualize the provided collection.

TODO:

explain the non obvious options (or all) here.

adjust drawing angle: If checked, the embeddings of the plabic graph (and tiling) will be rotated such that the boundary vertex 1 is drawn at a more consistant position.

file: saving, loading, export

edit