User guide

Introduction

We start off by giving some mathematical background, or rather by defining the needed language.

Weakly separated Collections

For any integer $n \geq 1$ we use the notation $[n]:=$ { $1, 2, \ldots, n$ } and denote by $\text{Pot}(k,n)$ the set of $k$-subsets of $[n]$.

Definition (weak separation)

Let $I, J$ be $k$-subsets of $[n]$, then we call $I$ and $J$ $\textbf{weakly separated}$ if we cannot find elements $a, c \in I \setminus J$ and $b, d \in J \setminus I$ such that $(a, b, c, d)$ is strictly cyclically ordered. In this case we write $I \parallel J$.

Intuitively two $k$-subsets are weakly separated if after can arranging $I \setminus J$ and $J \setminus I$ clockwise on a circle, they can be separated by a line.

Definition (weakly separated collection)

A subset $\mathcal{C} \subseteq \text{Pot}(k,n)$ is called a $\textbf{weakly separated collection}$ (abbreviated by WSC) if its elements are pairwise weakly separated. We often referr to elements of a WSC as labels.

Definition (mutation)

If $\mathcal{C}$ is a WSC that includes sets of the form $Iab, Ibc, Icd, Iad$ and $Iac$, where $(a,b,c,d)$ is strictly cyclically ordered. Then $\mathcal{C'} = (\mathcal{C} \setminus \{Iac\}) \cup \{Ibd\}$ is also a weakly separated collection, and we call the exchange of $Iac$ by $Ibd$ a $\textbf{mutation}$.

Plabic Tilings

Definition (plabic tiling)

Any WSC can be given the structure of an abstract $2$-dimensional cell complex, which in turn may be embedded into the plane. This construction will be called an (abstract) $\textbf{plabic tiling}$, and we referr to TODO for the mathematical details.

Intuitively a plabic tiling is a tiling of a convex $n$-gon into convex polygons, colored either black or white, and with vertices labelled by the elements of the underlying WSC. Plabic tilings are in bijective correspondance with WSC's that are maximal with respect to inclusion.

Plabic Graphs

Definition (plabic graph)

A $\textbf{plabic graph}$ is a finite simple connected plane graph $G$ whose interior is bounded by a vertex disjoint cycle containing $n$ $\textbf{boundary vertices}$ $b_1, \ldots, b_n$. Here the labelling is chosen in clockwise order.

We only consider $\textbf{reduced}$ plabic graphs which can be also seen to be in one to one correspondance to WSC's that are maximal with respect to inclusion. For more details we referr to TODO.

Creating a weakly separated collection

Vectors instead of sets

In this package we use vecors in place of sets, although WSC's are by definition sets of $k$-sets. We always assume such vectors to be increasingly ordered and not contain double elements. None of the below methods check these properties and unforseen behavior may arise if they are not fulfilled.

The data type for WSC's or rather (abstract) plabic tilings is given by WSCollection.

WeaklySeparatedCollections.WSCollection — Type

WSCollection

An abstract 2-dimensional cell complex living inside the matriod of k-sets in {1, ..., n}. Its vertices are labelled by elements of labels while quiver encodes adjacencies between the vertices. The 2-cells are colored black or white and contained in blackCliques and whiteCliques.

Optionally the 2-cells can be set to missing, to save memory.

Attributes

k::Int

n::Int

labels::Vector{Vector{Int}}

quiver::SimpleDiGraph{Int}

whiteCliques::Union{Missing, Dict{Vector{Int}, Vector{Int} }}

blackCliques::Union{Missing, Dict{Vector{Int}, Vector{Int} }}

Constructors

WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, computeCliques::Bool = true)
WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, quiver::SimpleDiGraph{Int}, 
                                                          computeCliques::Bool = true)

There are three different constructors to create a WSC:

WeaklySeparatedCollections.WSCollection — Method

WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, computeCliques::Bool = true)

Constructor of WSCollection. Adjacencies between its vertices as well as 2-cells are computed using only a set of vertex labels.

If computeCliques is set to false, the 2-cells will be set to missing.

WeaklySeparatedCollections.WSCollection — Method

WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, quiver::SimpleDiGraph{Int}, 
computeCliques::Bool = true)

Constructor of WSCollection. The 2-cells are computed from vertex labels as well as the their adjacencies encoded in quiver. Faster than just using labels most of the time.

If computeCliques is false the black and white 2-cells are set to missing instead.

WeaklySeparatedCollections.WSCollection — Method

WSCollection(collection::WSCollection; computeCliques::Bool = true)

Constructor of WSCollection. Computes 2-cells of collection if the are missing, othererwise returns a deepcopy of collection.

If computeCliques is false the black and white 2-cells are set to missing instead.

Thus to construct a WSC we only need to know its labels.

labels = [[1, 5, 6], [1, 2, 6], [1, 2, 3], [2, 3, 4], [3, 4, 5],
        [4, 5, 6], [2, 5, 6], [2, 3, 6], [3, 5, 6], [3, 4, 6]]
is_weakly_separated(6, labels)

true

C = WSCollection(3, 6, labels)

WSCollection of type (3,6) with 10 labels

However, if the underlying quiver is already known it can be passed to the constructor to speed up computations.

Q = C.quiver
WSCollection(3, 6, labels, Q)

WSCollection of type (3,6) with 10 labels

The last constructor is useful, if we already have a WSC but want to omit the 2-cells or if the 2-cells of our WSC are missing and we want to compute them.

D = WSCollection(C, computeCliques = false)
D.whiteCliques

missing

D = WSCollection(D)
D.whiteCliques

Dict{Vector{Int64}, Vector{Int64}} with 5 entries:
  [2, 3] => [3, 4, 8]
  [5, 6] => [1, 7, 9, 6]
  [2, 6] => [2, 8, 7]
  [3, 4] => [4, 5, 10]
  [3, 6] => [8, 10, 9]

Predefined collections

Although any WSC may be constructed as explained above, this can be quite tedious. Thus we provide shortcuts for the construction of some well known WSC's:

WeaklySeparatedCollections.checkboard_collection — Function

checkboard_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the checkboard graph.

WeaklySeparatedCollections.rectangle_collection — Function

rectangle_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the rectangle graph.

WeaklySeparatedCollections.dual_checkboard_collection — Function

dual_checkboard_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the dual-checkboard graph.

WeaklySeparatedCollections.dual_rectangle_collection — Function

dual_rectangle_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the dual-rectangle graph.

If we only want the underlying labels we may instead use

WeaklySeparatedCollections.rectangle_label — Function

rectangle_label(k::Int, n::Int, i::Int, j::Int)

Return the label of the rectangle graph in row i and column j.

WeaklySeparatedCollections.rectangle_labels — Function

rectangle_labels(k::Int, n::Int)

Return the labels of the rectangle graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.checkboard_label — Function

checkboard_label(k::Int, n::Int, i::Int, j::Int)

Return the label of the checkboard graph in row i and column j.

WeaklySeparatedCollections.checkboard_labels — Function

checkboard_labels(k::Int, n::Int)

Return the labels of the checkboard graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.dual_rectangle_label — Function

dual_rectangle_label(k::Int, n::Int, i::Int, j::Int)

Return the label of the dual rectangle graph in row i and column j.

WeaklySeparatedCollections.dual_rectangle_labels — Function

dual_rectangle_labels(k::Int, n::Int)

Return the labels of the dual-rectangle graph as a vector. The frozen labels are in positions 1 to n.

Missing docstring.

Missing docstring for dual_checkboard_label. Check Documenter's build log for details.

WeaklySeparatedCollections.dual_checkboard_labels — Function

dual_checkboard_labels(k::Int, n::Int)

Return the labels of the dual-checkboard graph as a vector. The frozen labels are in positions 1 to n.

Basic functionality

Armed with this plethora of examples, we are ready to discuss the basic functionalities of WSC's.

WSC's behave in many ways as their underlying arrays of labels would. In particular labels may be accessed directly.

rec = rectangle_collection(3, 6)
rec[3]

rec[7] = [1, 3, 6]

Caution is advised when modifying labels as above, as it is not checked if the resulting labels are still weakly separated nor is the associated data changed accordingly.

For convenience we also extend the following functions:

Base.in — Function

in(label::Vector{Int}, collection::WSCollection)

Return true label is occurs as label of collection.

Base.length — Function

length(collection::WSCollection)

Return the length of collection.labels.

Base.intersect — Function

intersect(collection1::WSCollection, collection2::WSCollection)

Return the common labels of collection1 and collection2.

Base.setdiff — Function

setdiff(collection1::WSCollection, collection2::WSCollection)

Return the labels of collection1 minus the labels of collection2.

Base.union — Function

setdiff(collection1::WSCollection, collection2::WSCollection)

Return the union of labels in collection1 and collection2.

Examples:

rec = rectangle_collection(3, 6)
check = checkboard_collection(3, 6)

check[10] in rec 

length(check)

intersect(rec, check) # similar for union and setdiff

Mutation

WSC's usually contain frozen elements that never change. On the other hand some elements may be modified via mutation and are called mutable. To figure out which elemnents of a WSC are frozen or mutable use the functions is_frozen or is_mutable.

WeaklySeparatedCollections.is_frozen — Function

is_frozen(collection::WSCollection, i::Int)

Return true if the vertex i of collection is frozen.

Examples

julia> H = rectangle_collection(4, 9)
julia> is_frozen(H, 5)
true

julia> is_frozen(H, 11)
false

WeaklySeparatedCollections.is_mutable — Function

is_mutable(collection::WSCollection, i::Int)

Return true if the vertex i of collection is mutable. This is the case if it is not frozen and is of degree 4.

Examples

julia> H = rectangle_collection(4, 9)
julia> is_mutable(H, 11)
false

julia> is_frozen(H, 10)
true

rec = rectangle_collection(3, 6)
is_frozen(rec, 4), is_mutable(rec, 7), is_mutable(rec, 11)

The frozen labels contained in any (maximal) WSC can be obtained via

WeaklySeparatedCollections.frozen_label — Function

frozen_label(k::Int, n::Int, i::Int)

return the i-th frozen label.

Examples

julia> frozen_label(3, 6, 3)
julia> [1,2,3]

WeaklySeparatedCollections.frozen_labels — Function

frozen_labels(k::Int, n::Int)

Return the frozen labels as a vector.

The indices of the mutable labels on a WSC can be obtained by using

WeaklySeparatedCollections.get_mutables — Function

get_mutables(collection::WSCollection)

Return all mutable vertices of collection.

Finally, to mutate a WSC, the functions mutateand mutate! are available.

WeaklySeparatedCollections.mutate! — Function

mutate!(collection::WSCollection, i::Int, mutateCliques::Bool = true)

Mutate the collection in direction i if i is a mutable vertex of collection.

If mutateCliques is set to false, the 2-cells are set to missing.

Examples

julia> H = rectangle_collection(4, 9)
julia> mutate!(H, 10)

mutate!(collection::WSCollection, label::Vector{Int}, mutateCliques::Bool = true)

Mutate the collection by addressing a vertex with its label.

Examples

julia> H = rectangle_collection(4, 9)
julia> H.labels[10]
4-element Vector{Int64}:
 2
 7
 8
 9

julia> mutate!(H, [2,7,8,9])

WeaklySeparatedCollections.mutate — Function

mutate(collection::WSCollection, i::Int, mutateCliques::Bool = true)

Version of mutate! that does not modify its arguments.

mutate(collection::WSCollection, label::Vector{Int}, mutateCliques::Bool = true)

Mutate the collection by addressing a vertex with its label, without modifying arguments.

Examples

other transformations

Apart from mutation, several other transformations of WSC's are available:

WeaklySeparatedCollections.rotate! — Function

rotate!(collection::WSCollection, amount::Int)

Rotate collection by amount, where a positive amount indicates a clockwise rotation.

Examples

julia> H = rectangle_collection(4, 9)
julia> rotate!(H, 2)

WeaklySeparatedCollections.rotate — Function

rotate(collection::WSCollection, amount::Int)

Version of rotate! that does not modify its argument.

WeaklySeparatedCollections.reflect! — Function

reflect!(collection::WSCollection, axis::Int = 1)

Reflect collection by letting the permutation x ↦ 2*axis -x interpreted modulo n = collection.n act on the labels of collection.

Examples

julia> H = rectangle_collection(4, 9)
julia> reflect!(H, 1)

WeaklySeparatedCollections.reflect — Function

reflect(collection::WSCollection, axis::Int = 1)

Version of reflect! that does not modify its argument.

WeaklySeparatedCollections.complement! — Function

complement!(collection::WSCollection)

Return the collection whose labels are complementary to those of collection.

Examples

julia> H = rectangle_collection(4, 9)
julia> complement!(H)

Missing docstring.

Missing docstring for complement. Check Documenter's build log for details.

WeaklySeparatedCollections.swap_colors! — Function

swap_colors!(collection::WSCollection)

Return the weakly separated collection whose corresponding plabic graph is obtained from the one of collection by swapping the colors black and white.

This is the same as taking complements and rotating by collection.k.

Examples

julia> H = rectangle_collection(4, 9)
julia> swap_colors!(H)

WeaklySeparatedCollections.swap_colors — Function

swap_colors(collection::WSCollection)

Version of swap_colors! that does not modify its argument.

We often want to deal with maximal WSC's instead of their subsets. To extend a given WSC to a maximal one, the following functions may be used:

WeaklySeparatedCollections.extend_weakly_separated! — Function

extend_weakly_separated!(k::Int, n::Int, labels::Vector{Vector{Int}})

Extend labels to contain the labels of a maximal weakly separated collection.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> extend_weakly_separated!(4, 9, labels)

extend_weakly_separated!(k::Int, n::Int, labels1::Vector{Vector{Int}}, 
                                         labels2::Vector{Vector{Int}})

Extend labels1 to contain the labels of a maximal weakly separated collection. Use elements of labels2 if possible.

Examples

julia> labels1 = [[1,3,5,7], [1,3,5,9]]
julia> labels2 = checkboard_labels(4, 9)
julia> extend_weakly_separated!(4, 9, labels1, labels2)

extend_weakly_separated!(labels::Vector{Vector{Int}}, collection::WSCollection)

Extend labels to contain the labels of a maximal weakly separated collection. Use labels of collection if possible.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> H = checkboard_collection(4, 9)
julia> extend_weakly_separated!(labels, H)

WeaklySeparatedCollections.extend_to_collection — Function

extend_to_collection(k::Int, n::Int, labels::Vector{Vector{Int}})

Return a maximal weakly separated collection containing all elements of labels.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> extend_to_collection(4, 9, labels)

extend_to_collection(k::Int, n::Int, labels1::Vector{Vector{Int}}, 
                                     labels2::Vector{Vector{Int}})

Return a maximal weakly separated collection containing all elements of labels1. Use elements of labels2 if possible.

Examples

julia> labels1 = [[1,3,5,7], [1,3,5,9]]
julia> labels2 = checkboard_labels(4, 9)
julia> extend_to_collection(4, 9, labels1, labels2)

extend_to_collection(labels::Vector{Vector{Int}}, collection::WSCollection)

Return a maximal weakly separated collection containing all elements of labels. Use labels of collection if possible.

Examples

julia> labels = [[1,3,5,7], [1,3,5,9]]
julia> H = checkboard_collection(4, 9)
julia> extend_to_collection(labels, H)

Plotting

Plotting WSC's requires Luxor to be installed and loaded as detailed here.

In the introduction we learned about plabic tilings as well as plabic graphs as objects living in the plane which are in one to one correspndance to maximal WSC's. Thus we can plot a maximal WSC using its corresponding plabic tiling or plabic graph. The functions to accomplish this are:

WeaklySeparatedCollections.drawTiling — Function

drawTiling(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic tiling of the provided weakly separated collection and save it as an image file of specified size. Both the name as well as the resulting file type of the image are controlled via title.

Inside a Jupyter sheet drawing without saving an image is possible by omitting the title argument i.e. via drawTiling(collection::WSCollection, width::Int = 500, height::Int = 500).

Keyword Arguments

topLabel = nothing

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

highlightMutables::Bool = true

labelDirection = "left"

toplabel controls the rotation of the drawing by drawing the specified label at the top. labelDirection controls whether the "left" (i.e. the usual ones) or "right" (complements) labels are drawn.

WeaklySeparatedCollections.drawPLG — Function

drawPLG(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic graph of the provided weakly separated collection and save it as an image file of specified size. Both the name as well as the resulting file type of the image are controlled via title.

Inside a Jupyter sheet drawing without saving an image is possible by omitting the title argument i.e. via drawPLG(collection::WSCollection, width::Int = 500, height::Int = 500).

Keyword Arguments

topLabel = nothing

drawmode::String = "straight"

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

highlightMutables::Bool = false

labelDirection = "left"

toplabel controls the rotation of the drawing by drawing the specified label at the top. drawmode controls how edges are drawn and may be choosen as "straight", "smooth" or "polygonal". labelDirection controls whether the "left" (i.e. the usual ones) or "right" (complements) labels are drawn.

Graphical user interface

This section is work in progress.

The graphical user interface requires both an installation of Luxor as well as Mousetrap. See here for details.

While plotting WSC's enables us to visualize them, the resulting images lack interactivity. This is where the built in gui application comes in handy. To start it we use

WeaklySeparatedCollections.visualizer! — Function

visualizer!(collection::WSCollection = rectangle_collection(4, 9))

Start the graphical user interface to visualize the provided collection.

TODO:

WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}; computeCliques::Bool = true)

explain the non obvious options (or all) here.

adjust drawing angle: If checked, the embeddings of the plabic graph (and tiling) will be rotated such that the boundary vertex 1 is drawn at a more consistant position.

file: saving, loading, export

edit