User guide

Combinatorics

Hier könnte eine eloquente Einleitung stehen.

WeaklySeparatedCollections.is_weakly_separated — Function

is_weakly_separated(n::Int, v::Vector{Int}, w::Vector{Int})

Test if two vectors v and w viewed as subsets of {1 , ..., n } are weakly separated.

is_weakly_separated(n::Int, labels::Vector{Vector{Int}})

Test if the vectors contained in labels are pairwise weakly separated.

WeaklySeparatedCollections.rectangle_labels — Function

rectangle_labels(k::Int, n::Int)

Return the labels of the rectangle graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.checkboard_labels — Function

checkboard_labels(k::Int, n::Int)

Return the labels of the checkboard graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.dual_rectangle_labels — Function

dual_rectangle_labels(k::Int, n::Int)

Return the labels of the dual-rectangle graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.dual_checkboard_labels — Function

dual_checkboard_labels(k::Int, n::Int)

Return the labels of the dual-checkboard graph as a vector. The frozen labels are in positions 1 to n.

WeaklySeparatedCollections.WSCollection — Type

WSCollection

An abstract 2-dimensional cell complex living inside the matriod of k-sets in {1, ..., n}. Its vertices are labelled by elements of labels while quiver encodes adjacencies between the vertices. The 2-cells are colored black or white and contained in blackCliques and whiteCliques.

Optionally the 2-cells can be set to missing, to save memory.

Attributes

k::Int

n::Int

labels::Vector{Vector{Int}}

quiver::SimpleDiGraph{Int}

whiteCliques::Union{Missing, Dict{Vector{Int}, Vector{Int} }}

blackCliques::Union{Missing, Dict{Vector{Int}, Vector{Int} }}

Constructors

WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, computeCliques::Bool = true)
WSCollection(k::Int, n::Int, labels::Vector{Vector{Int}}, quiver::SimpleDiGraph{Int}, 
                                                          computeCliques::Bool = true)

Base.isequal — Function

isequal(collection1::WSCollection, collection2::WSCollection)

Return true if the vertices of collection1 and collection2 share the same labels. The order of labels in each collection does not matter.

Base.:== — Function

(==)(collection1::WSCollection, collection2::WSCollection)

Return true if the vertices of collection1 and collection2 share the same labels. The order of labels in each collection does not matter.

WeaklySeparatedCollections.is_frozen — Function

is_frozen(collection::WSCollection, i::Int)

Return true if the vertex i of collection is frozen.

WeaklySeparatedCollections.is_mutable — Function

is_mutable(collection::WSCollection, i::Int)

Return true if the vertex i of collection is mutable. This is the case if it is not frozen and is of degree 4.

WeaklySeparatedCollections.mutate! — Function

mutate!(collection::WSCollection, i::Int, mutateCliques::Bool = true)

Mutate the collection in direction i if i is a mutable vertex of collection.

If mutateCliques is set to false, the 2-cells are set to missing.

mutate!(collection::WSCollection, label::Vector{Int}, mutateCliques::Bool = true)

Mutate the collection by addressing a vertex with its label.

WeaklySeparatedCollections.mutate — Function

mutate(collection::WSCollection, i::Int, mutateCliques::Bool = true)

Version of mutate! that does not modify its arguments.

mutate(collection::WSCollection, label::Vector{Int}, mutateCliques::Bool = true)

Mutate the collection by addressing a vertex with its label, without modifying arguments.

WeaklySeparatedCollections.checkboard_collection — Function

checkboard_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the checkboard graph.

WeaklySeparatedCollections.rectangle_collection — Function

rectangle_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the rectangle graph.

WeaklySeparatedCollections.dual_checkboard_collection — Function

dual_checkboard_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the dual-checkboard graph.

WeaklySeparatedCollections.dual_rectangle_collection — Function

dual_rectangle_collection(k::Int, n::Int)

Return the weakly separated collection corresponding to the dual-rectangle graph.

WeaklySeparatedCollections.rotate_collection! — Function

rotate_collection!(collection::WSCollection, amount::Int)

Rotate collection by amount, where a positive amount indicates a clockwise rotation.

WeaklySeparatedCollections.rotate_collection — Function

rotate_collection(collection::WSCollection, amount::Int)

Version of rotate_collection! that does not modify its argument.

WeaklySeparatedCollections.reflect_collection! — Function

reflect_collection!(collection::WSCollection, axis::Int = 1)

Reflect collection by letting the permutation x ↦ 2*axis -x interpreted modulo n = collection.n act on the labels of collection.

WeaklySeparatedCollections.reflect_collection — Function

reflect_collection(collection::WSCollection, axis::Int = 1)

Version of reflect_collection! that does not modify its argument.

WeaklySeparatedCollections.complement_collection! — Function

complement_collection!(collection::WSCollection)

Return the collection whose labels are complementary to those of collection.

WeaklySeparatedCollections.complement_collection — Function

complement_collection(collection::WSCollection)

Version of complement_collection! that does not modify its argument.

WeaklySeparatedCollections.swaped_colors_collection! — Function

swaped_colors_collection!(collection::WSCollection)

Return the weakly separated collection whose corresponding plabic graph is obtained from the one of collection by swapping the colors black and white.

This is the same as taking complements and rotating by collection.k.

WeaklySeparatedCollections.swaped_colors_collection — Function

swaped_colors_collection(collection::WSCollection)

Version of swaped_colors_collection! that does not modify its argument.

WeaklySeparatedCollections.extend_weakly_separated! — Function

extend_weakly_separated!(k::Int, n::Int, labels::Vector{Vector{Int}})

Extend labels to contain the labels of a maximal weakly separated collection.

extend_weakly_separated!(k::Int, n::Int, labels1::Vector{Vector{Int}}, 
                                         labels2::Vector{Vector{Int}})

Extend labels1 to contain the labels of a maximal weakly separated collection. Use elements of labels2 if possible.

extend_weakly_separated!(labels::Vector{Vector{Int}}, collection::WSCollection)

Extend labels to contain the labels of a maximal weakly separated collection. Use labels of collection if possible.

WeaklySeparatedCollections.extend_to_collection — Function

extend_to_collection(k::Int, n::Int, labels::Vector{Vector{Int}})

Return a maximal weakly separated collection containing all elements of labels.

extend_to_collection(k::Int, n::Int, labels1::Vector{Vector{Int}}, 
                                     labels2::Vector{Vector{Int}})

Return a maximal weakly separated collection containing all elements of labels1. Use elements of labels2 if possible.

extend_to_collection(labels::Vector{Vector{Int}}, collection::WSCollection)

Return a maximal weakly separated collection containing all elements of labels. Use labels of collection if possible.

<!– TODO computecliques, computeadjacencies, computeboundaries, superpotential_labels –>

Plotting

WeaklySeparatedCollections.drawTiling — Function

drawTiling(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic tiling of the provided weakly separated collection and save it as an image file of specified size. Both the name as well as the resulting file type of the image are controlled via title.

Keyword Arguments

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

adjustAngle::Bool = false

highlightMutables::Bool = true

scale::Float64 = 0.0

Examples

WeaklySeparatedCollections.drawPLG_straight — Function

drawPLG_straight(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic graph of the provided weakly separated collection and save it as an image file of specified size. All inner edges are drawn as straight line. Both the name as well as the resulting file type of the image are controlled via title.

Keyword Arguments

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

adjustAngle::Bool = false

highlightMutables::Bool = true

scale::Float64 = 0.0

Examples

WeaklySeparatedCollections.drawPLG_smooth — Function

drawPLG_smooth(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic graph of the provided weakly separated collection and save it as an image file of specified size. All inner edges are drawn as smooth curves. Both the name as well as the resulting file type of the image are controlled via title.

Keyword Arguments

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

adjustAngle::Bool = false

scale::Float64 = 0.0

Examples

WeaklySeparatedCollections.drawPLG_poly — Function

drawPLG_smooth(collection::WSCollection, title::String, width::Int = 500, height::Int = 500)

Draw the plabic graph of the provided weakly separated collection and save it as an image file of specified size. All inner edges are drawn as polygonal curves. Both the name as well as the resulting file type of the image are controlled via title.

Keyword Arguments

backgroundColor::Union{String, ColorTypes.Colorant} = ""

drawLabels::Bool = true

adjustAngle::Bool = false

scale::Float64 = 0.0

Examples

Graphical user interface

WeaklySeparatedCollections.visualizer! — Function

visualizer!(collection::WSCollection = rectangle_collection(4, 9))

Start the graphical user interface to visualize the provided collection.